backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
bcrypt
pydantic
